move AuthenticationManagerResolver bean to OAuth2BeansConfig
